+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] 
 
+## [4.4.1](v4.4.1) - 2024-07-29 14:21:42+02:00 (urgency: medium)
+* fixed `hal_py.GateLibrary.gate_types` pybind
+* fixed `hal_py.NetlistFactory.load_netlist` pybind
+
 ## [4.4.0](v4.4.0) - 2024-07-19 15:55:24+02:00 (urgency: medium)
 * **WARNING:** this release breaks the API of the `boolean_influence` and `bitorder_propagation` plugin
 * **WARNING:** this release contains many new unstable plugin APIs that will likely change in the future
@@ -33,6 +37,7 @@ All notable changes to this project will be documented in this file.
     * added `json.hpp` from nlohmann to deps to offer a light weight json api
     * adapted cmake to consider the correct flags when finding and linking against the new version of Bitwuzla
 * miscellaneous
+  * added logic evaluator plugin
   * added backward compatibility for view management
   * slightly improved symbolic execution engine
   * added a version of `netlist_factory::load_netlist` that takes a path to a netlist file as well as a pointer to a gate library